fix: move utils to Complexes
